+++ business/src/main/java/com/kevinguanchedarias/owgejava/repository/PlanetListRepository.java
@@ -1,0 +1,27 @@
+package com.kevinguanchedarias.owgejava.repository;
+
+import java.util.List;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.kevinguanchedarias.owgejava.entity.PlanetList;
+import com.kevinguanchedarias.owgejava.entity.embeddedid.PlanetUser;
+
+/**
+ *
+ * @since 0.9.0
+ * @author Kevin Guanche Darias <kevin@kevinguanchedarias.com>
+ *
+ */
+public interface PlanetListRepository extends JpaRepository<PlanetList, PlanetUser> {
+
+	/**
+	 *
+	 * @param userId
+	 * @return
+	 * @since 0.9.0
+	 * @author Kevin Guanche Darias <kevin@kevinguanchedarias.com>
+	 */
+	List<PlanetList> findByPlanetUserUserId(Integer userId);
+
+}