+++ game-frontend/projects/owge-universe/src/lib/types/planet-list-item.type.ts
@@ -1,0 +1,15 @@
+import { Planet } from '../pojos/planet.pojo';
+
+/**
+ * Represents an item of the uset planet list
+ *
+ * @author Kevin Guanche Darias <kevin@kevinguanchedarias.com>
+ * @since 0.9.0
+ * @export
+ */
+export interface PlanetListItem {
+    userId: number;
+    username: string;
+    planet: Planet;
+    name: string;
+}